leados-agent-example/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
rapidfuzz==3.9.0
fake-useragent==1.5.1
google-generativeai==0.8.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.23.8
anyio==4.4.0
lxml==5.2.2
sib-api-v3-sdk==7.6.0
supabase==2.7.0
python-telegram-bot==21.5
aiofiles==23.2.1

openpyxl>=3.1.2
xlrd>=2.0.1
python-multipart>=0.0.9
slowapi==0.1.9
groq==0.9.0
